chore(SharedModule) rename file for SharedModule
modules/shared/module.ts -> modules/shared/shared.module.ts
