+++ web-ui/src/app/modules/auth/auth.module.ts
@@ -5,7 +5,7 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { MatExpansionModule, MatFormFieldModule, MatInputModule, MatStepperModule } from '@angular/material';
 import { RouterModule } from '@angular/router';
 import { AccountModule } from 'app/modules/account/account.module';
-import { SharedModule } from 'app/modules/shared/module';
+import { SharedModule } from 'app/modules/shared/shared.module';
 
 import { AuthLoginFormComponent } from './components/auth-login-form/auth-login-form.component';
 import { AuthRegisterFormComponent } from './components/auth-register-form/auth-register-form.component';
