+++ web-ui/src/app/modules/exchanges/exchanges.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MatChipsModule, MatExpansionModule, MatFormFieldModule, MatInputModule, MatSelectModule, MatTableModule } from '@angular/material';
-import { SharedModule } from 'app/modules/shared/module';
+import { SharedModule } from 'app/modules/shared/shared.module';
 
 import { ExchangeConfigurationItemComponent } from './components/exchange-configuration-item/exchange-configuration-item.component';
 import { ExchangesConfigurationListComponent } from './components/exchange-configuration-list/exchange-configuration-list.component';
