+++ web-ui/src/app/modules/home/home.module.ts
@@ -1,6 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { SharedModule } from 'app/modules/shared/module';
+import { SharedModule } from 'app/modules/shared/shared.module';
 
 import { HomePage } from './pages/home/home.page';
 import { HomeRoutingModule } from './home-routing.module';
