+++ web-ui/src/app/modules/reports/reports.module.ts
@@ -1,6 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { SharedModule } from 'app/modules/shared/module';
+import { SharedModule } from 'app/modules/shared/shared.module';
 
 import { ReportsPage } from './pages/reports/reports.page';
 import { ReportsRoutingModule } from './reports-routing.module';
